Focus First Name when a new contact is created
⌘N (and the ＋ button) created a contact and selected it, but focus
stayed nowhere — you had to click into First Name before typing.
ContentView now records the new contact's id and passes a one-shot
`focusNameField` flag to ContactDetailView, which focuses the First Name
field on appear and clears the flag (so reselecting an existing contact
doesn't steal focus). The flag is keyed to the created id, and the detail
view is already `.id`-recreated per contact, so only the new one focuses.

To stay under the file/type-body length limits, moved the photo
import/remove handlers into a new ContactDetailView+Photo.swift extension
(errorMessage/store relaxed from private to internal so the extension can
reach them).

Tests: none new — focus-on-appear is view behavior with no extractable
logic. 119 unit tests green; build/lint/format clean.

Entire-Checkpoint: c619f5f46467
